Modify: dto 타입인 것과 아닌 것을 다른 파일로 분리
