+++ src/types/tteokguk.dto.ts
@@ -1,0 +1,16 @@
+import { IngredientKey } from "./ingredient";
+
+export interface PostTteokgukRequest {
+  wish: string;
+  ingredients: IngredientKey[];
+  access: boolean;
+}
+
+export interface PostTteokgukResponse {
+  tteokgukId: number;
+  memberId: number;
+  wish: string;
+  access: boolean;
+  completion: boolean;
+  ingredients: IngredientKey[];
+}
